backend/requirements.txt

Flask
flask
flask_cors
pymongo
werkzeug
python-dotenv
email_validator
flask_jwt_extended
timedelta
datetime
pytesseract
pdfplumber
python-docx
cloudinary


